Collections

Nakala.Collections.deletecollections — Function

deletecollections(identifier::String, headers::Dict, apiTest=false)

Suppression d'une collection.

exemple

Nakala.Collections.deletecollections_datas — Function

deletecollections_datas(identifier::String, headers::Dict, body::Array, apiTest::Bool=false)

Suppression d'une liste de données d'une collection.

exemple

Nakala.Collections.deletecollections_metadatas — Function

deletecollections_metadatas(identifier::String, headers::Dict, body::Dict, apiTest::Bool=false)

Suppression de métadonnées pour une collection.

exemple

Nakala.Collections.deletecollections_rights — Function

deletecollections_rights(identifier::String, headers::Dict, body::Dict, apiTest::Bool=false)

Suppression des droits pour utilisateur ou un groupe d'utilisateurs sur une collection.

exemple

Nakala.Collections.getcollections — Function

getcollections(identifier::String, headers::Dict, apiTest=false)

Récupération des informations sur une collection.

exemple

julia> getcollections(identifier, headers, true)
Dict{String, Any} with 3 entries:
  "body"      => Dict{String, Any}("haveAccessibleData"=>false, "websitePrefix"=>"", "isDepositor"=>true, "isOwner"=>true, "depositor"=>Dic…
  "status"    => 200
  "isSuccess" => true

Nakala.Collections.getcollections_datas — Function

getcollections_datas(identifier::String, params::Array, headers::Dict, apiTest::Bool=false)

Récupération de la liste paginée des données contenues dans la collection.

exemple

Nakala.Collections.getcollections_metadatas — Function

getcollections_metadatas(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération des métadonnées d'une collection.

exemple

Nakala.Collections.getcollections_rights — Function

getcollections_rights(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération des utilisateurs et des groupes ayant des droits sur la collection.

exemple

Nakala.Collections.getcollections_status — Function

getcollections_status(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération du statut d'une collection.

exemple

Nakala.Collections.postcollections — Function

postcollections(headers::Dict, body::Dict, apiTest::Bool=false)

Création d'une nouvelle collection.

exemple

Nakala.Collections.postcollections_datas — Function

postcollections_datas(identifier::String, headers::Dict, body::Array, apiTest::Bool=false)

Ajout d'une liste de données dans une collection.

exemple

Nakala.Collections.postcollections_metadatas — Function

postcollections_metadatas(identifier::String, headers::Dict, body::Dict, apiTest::Bool=false)

Ajout d'une nouvelle métadonnée à une collection.

exemple

Nakala.Collections.postcollections_rights — Function

postcollections_rights(identifier::String, headers::Dict, body::Array, apiTest::Bool=false)

Ajout de droits sur une collection.

exemple

Nakala.Collections.putcollections — Function

putcollections(identifier::String, headers::Dict, body::Dict, apiTest=false)

Modification des informations d'une collection.

exemple

Nakala.Collections.putcollections_status — Function

putcollections_status(identifier::String, status::String, headers::Dict, apiTest::Bool=false)

Changement du statut d'une collection.

exemple