Datas

Le module Nakala.Datas correspond au end-point Nakala datas. Il est dédié à la gestion des données.

Nakala.Datas.deletedatas — Method

deletedatas(identifier::String, headers::Dict; apitest::Bool=false)

Suppression d'une donnée.

exemple

Nakala.Datas.deletedatas_collections — Method

deletedatas_collections(identifier::String, headers::Dict, body::Array; apitest::Bool=false)

Suppression d'une donnée d'un ensemble de collections.

exemple

Nakala.Datas.deletedatas_files — Method

deletedatas_files(identifier::String, fileIdentifier::String, headers::Dict; apitest::Bool=false)

Supprime un fichier (argument fileIdentifier) de la donnée Nakala dont l'identifiant correspond à identifier.

exemple

julia> Nakala.Datas.deletedatas_files("10.34847/nkl.b0ee", "a0b65939670bc2c01", headers, apitest=true)
Dict{String, Any} with 3 entries:
  "body"      => ""
  "status"    => 204
  "isSuccess" => true

Nakala.Datas.deletedatas_metadatas — Method

deletedatas_metadatas(identifier::String, headers::Dict, body::Dict; apitest::Bool=false)

Suppression de métadonnées pour une donnée.

exemple

Nakala.Datas.deletedatas_relations — Method

deletedatas_relations(identifier::String, headers::Dict, body::Dict; apitest::Bool=false)

Suppression des relations sur une donnée.

exemple

Nakala.Datas.deletedatas_rights — Method

deletedatas_rights(identifier::String, headers::Dict, body::Dict; apitest::Bool=false)

Suppression des droits pour un utilisateur ou un groupe d'utilisateurs sur une donnée.

exemple

Nakala.Datas.deletedatas_uploads — Method

deletedatas_uploads(fileIdentifier::String, headers::Dict; apitest::Bool=false)

Suppression d'un fichier déposé dans l'espace temporaire

exemple

Nakala.Datas.downloaddatas_files — Method

downloaddatas_files(identifier::String, outputDir::String, header::Dict; apitest::Bool=false)

Téléchargement des fichiers déposés associés à une donnée.

exemple

Nakala.Datas.getdatas — Method

getdatas(identifier::String, headers::Dict; apitest::Bool=false)

Récupère les informations relatives à une donnée aynt pour identifiant identifier.

exemple

julia> headers = Dict( "X-API-KEY" => "01234567-89ab-cdef-0123", :accept => "application/json" )
Dict{Any, String} with 2 entries:
  :accept     => "application/json"
  "X-API-KEY" => "01234567-89ab-cdef-0123"

julia> Nakala.Datas.getdatas("10.34847/nkl.af219", headers)
Dict{String, Any} with 3 entries:
  "body"      => Dict{String, Any}("isDepositor"=>false, "isOwner"=>false, "depositor"=>Dict{String, Any}("name"=>"…,…", "photo"=>nothing, "surname"=>"…
  "status"    => 200
  "isSuccess" => true

Nakala.Datas.getdatas_collections — Method

getdatas_collections(identifier::String, headers::Dict; apitest::Bool=false)

Récupération de la liste des collections contenant la donnée.

exemple

Nakala.Datas.getdatas_files — Method

getdatas_files(identifier::String, headers::Dict; apitest::Bool=false)

Récupération des métadonnées des fichiers associés à une donnée.

exemple

Nakala.Datas.getdatas_metadatas — Method

getdatas_metadatas(identifier::String, headers::Dict; apitest::Bool=false)

Récupération de la liste des métadonnées d'une donnée.

exemple

Nakala.Datas.getdatas_relations — Method

getdatas_relations(identifier::String, headers::Dict; apitest::Bool=false)

Récupération de la liste des relations d'une donnée.

exemple

Nakala.Datas.getdatas_rights — Method

getdatas_rights(identifier::String, headers::Dict; apitest::Bool=false)

Récupération des groupes et des utilisateurs ayant des droits sur la donnée.

exemple

Nakala.Datas.getdatas_status — Method

getdatas_status(identifier::String, headers::Dict; apitest::Bool=false)

Récupération du statut d'une donnée.

exemple

Nakala.Datas.getdatas_uploads — Method

getdatas_uploads(headers::Dict; apitest::Bool=false)

Récupération pour un utilisateur de la liste des objets fichiers déposés non encore associés à une donnée.

exemple

Nakala.Datas.getdatas_version — Method

getdatas_version(identifier::String, headers::Dict; apitest::Bool=false)

Récupère la liste des versions d'une donnée.

exemple

julia> headers = Dict( "X-API-KEY" => "123456789-abcdefghij", :accept => "application/json" )
Dict{Any, String} with 2 entries:
  :accept     => "application/json"
  "X-API-KEY" => "123456789-abcdefghij"

julia> Nakala.Datas.getdatas_version("10.34847/nkl.93aeeeee", headers)
Dict{String, Any} with 3 entries:
  "body"      => Dict{String, Any}("currentPage"=>1, "lastPage"=>1, "total"=>1, "data"=>Any[Dict{String, Any}("modDate"=>"2023-04-23T20:24:49+02:00", "versionIdentifie…
  "status"    => 200
  "isSuccess" => true

Nakala.Datas.postdatas — Method

postdatas(headers::Dict, body::Dict; apitest::Bool=false)

Dépose une nouvelle donnée sur Nakala. Les métadonnées nécessaires à la création de la donnée sont déclarées avec l'argument body

exemple

julia> headers = Dict( "X-API-KEY" => "01234567-89ab-cdef-0123", :accept => "application/json" )
Dict{Any, String} with 2 entries:
  :accept     => "application/json"
  "X-API-KEY" => "01234567-89ab-cdef-0123"

julia> body = Dict(
  :collectionsIds => [],
  :files => [ Dict("name" => "file.txt", "sha1" => sha1, "embargoed" => "2024-09-01") ],
  :status => "pending",
  :metas => [
    Dict(:value => "Ma données", :propertyUri => "http://nakala.fr/terms#title", :lang => "fr", :typeUri => "http://www.w3.org/2001/XMLSchema#string"),
    Dict(:value => "http://purl.org/coar/resource_type/c_18cf", :propertyUri => "http://nakala.fr/terms#type", :typeUri => "http://www.w3.org/2001/XMLSchema#anyURI"),
    Dict(:value => Dict(:surname => "Rémi", :givenname => "Fassol"), :propertyUri => "http://nakala.fr/terms#creator", :typeUri => "http://www.w3.org/2001/XMLSchema#string"),
    Dict(:value => "2024-09-01", :propertyUri => "http://nakala.fr/terms#created", :typeUri => "http://www.w3.org/2001/XMLSchema#string"),
    Dict(:value => "PDM", :propertyUri => "http://nakala.fr/terms#license", :typeUri => "http://www.w3.org/2001/XMLSchema#string"),
    Dict(:value => "Description", :propertyUri => "http://purl.org/dc/terms/description", :lang => "fr", :typeUri => "http://www.w3.org/2001/XMLSchema#string"),
    Dict( :value => Dict( :surname => "Rémi", :givenname => "Fassol" ), :propertyUri => "http://nakala.fr/terms#creator" )
  ],
  :rights => []
)

julia> Nakala.Datas.postdatas(headers, body, apitest=true)
Dict{String, Any} with 3 entries:
  "body"      => Dict{String, Any}("message"=>"Data created", "payload"=>Dict{String, Any}("id"=>"10.34847/nkl.35b5"), "code"=>201)
  "status"    => 201
  "isSuccess" => true

Nakala.Datas.postdatas_collections — Method

postdatas_collections(identifier::String, headers::Dict, body::Array; apitest::Bool=false)

Ajout d'une donnée dans un ensemble de collections.

exemple

Nakala.Datas.postdatas_files — Method

postdatas_files(identifier::String, headers::Dict, body::Dict; apitest::Bool=false)

Ajoute un fichier à une donnée Nakala dont l'identifiant correspond à l'argument identifier. Au préalable, le fichier doit avoir été déposé sur l'espace temporaire de Nakala (postdatas_uploads()) et vous devez disposer de son identifiant sha1.

exemple

julia> headers = Dict( "X-API-KEY" => "01234567-89ab-cdef-0123", :accept => "application/json" )
Dict{Any, String} with 2 entries:
  :accept     => "application/json"
  "X-API-KEY" => "01234567-89ab-cdef-0123"

julia> body = Dict( "description" => "Greetings.", "sha1" => sha1, "embargoed" => "2024-09-01" )
Dict{String, String} with 3 entries:
  "embargoed"   => "2024-09-01"
  "sha1"        => "49aaa64346f03598951ad578"
  "description" => "Greetings."

julia> Nakala.Datas.postdatas_files("10.34847/nkl.c61ej8x9", headers, body, apitest=true)
Dict{String, Any} with 3 entries:
  "body"      => Dict{String, Any}("message"=>"File added", "code"=>200)
  "status"    => 200
  "isSuccess" => true

Nakala.Datas.postdatas_metadatas — Method

postdatas_metadatas(identifier::String, headers::Dict, body::Dict; apitest::Bool=false)

Suppression de métadonnées pour une donnée.

exemple

Nakala.Datas.postdatas_relations — Method

postdatas_relations(identifier::String, headers::Dict, body::Array; apitest::Bool=false)

Ajout de relations sur une donnée.

exemple

Nakala.Datas.postdatas_rights — Method

postdatas_rights(identifier::String, headers::Dict, body::Array; apitest::Bool=false)

Ajout de droits sur une donnée.

exemple

Nakala.Datas.postdatas_uploads — Method

postdatas_uploads(file::String, headers::Dict; apitest::Bool=false)

Dépose un fichier sur l'espace temporaire de Nakala et retourne un dictionnaire. La réponse du serveur correspond à la valeur de la clé body.

exemple

julia> headers = Dict( "X-API-KEY" => "01234567-89ab-cdef-0123", :accept => "application/json" )
Dict{Any, String} with 2 entries:
  :accept     => "application/json"
  "X-API-KEY" => "01234567-89ab-cdef-0123"

julia> Nakala.Datas.postdatas_uploads("path/to/file.txt", headers, apitest=true)
Dict{String, Any} with 3 entries:
  "body"      => Dict{String, Any}("name"=>"file.txt", "sha1"=>"a0b65939670bc2c010f4d5d6a0b3e")
  "status"    => 201
  "isSuccess" => true

Nakala.Datas.putdatas — Method

putdatas(identifier::String, headers::Dict, body::Dict; apitest::Bool=false)

Modifie informations d'une donnée.

exemple

julia> new_metadata = Dict(
  "metas" => [
    Dict("value" => "My data", "propertyUri" => "http://nakala.fr/terms#title", "lang" => "en", "typeUri" => "http://www.w3.org/2001/XMLSchema#string"),
    Dict("value" => "http://purl.org/coar/resource_type/c_18cf", "propertyUri" => "http://nakala.fr/terms#type", "typeUri" => "http://www.w3.org/2001/XMLSchema#anyURI"),
    Dict(:value => Dict(:surname => "Rémi", :givenname => "Fassol"), :propertyUri => "http://nakala.fr/terms#creator", :typeUri => "http://www.w3.org/2001/XMLSchema#string"),
    Dict("value" => "2024-09-01", "propertyUri" => "http://nakala.fr/terms#created", "typeUri" => "http://www.w3.org/2001/XMLSchema#string"),
    Dict("value" => "PDM", "propertyUri" => "http://nakala.fr/terms#license", "typeUri" => "http://www.w3.org/2001/XMLSchema#string"),
    Dict("value" => "New description.", "propertyUri" => "http://purl.org/dc/terms/description", "lang" => "en", "typeUri" => "http://www.w3.org/2001/XMLSchema#string")
  ]
)
Dict{String, Vector{Dict}} with 1 entry:
"metas" => [Dict("typeUri"=>"http://www.w3.org/2001/XMLSchema#string", "propertyUri"=>"http://nakala.fr/terms#title", "lang"=>"en", "value"=>"My data"), Dict("typeUr…

julia> Nakala.Datas.putdatas("10.34847/nkl.83409aze", headers, new_metadata, apitest=true)
Dict{String, Any} with 3 entries:
  "body"      => ""
  "status"    => 204
  "isSuccess" => true

Nakala.Datas.putdatas_collections — Method

putdatas_collections(identifier::String, headers::Dict, body::Array; apitest::Bool=false)

Remplacement de l'ensemble des collections d'une donnée.

exemple

Nakala.Datas.putdatas_status — Method

putdatas_status(identifier::String, status::String, headers::Dict; apitest::Bool=false)

Changement du statut d'une donnée.

exemple