Datas

Nakala.Datas.deletedatas — Function

deletedatas(identifier::String, headers::Dict, apiTest=false)

Suppression d'une donnée.

exemple

Nakala.Datas.deletedatas_collections — Function

deletedatas_collections(identifier::String, headers::Dict, body::Array, apiTest::Bool=false)

Suppression d'une donnée d'un ensemble de collections.

exemple

Nakala.Datas.deletedatas_files — Function

deletedatas_files(identifier::String, fileIdentifier::String, headers::Dict, apiTest::Bool=false)

Suppression de fichier à une donnée.

exemple

Nakala.Datas.deletedatas_metadatas — Function

deletedatas_metadatas(identifier::String, headers::Dict, body::Dict, apiTest::Bool=false)

Suppression de métadonnées pour une donnée.

exemple

Nakala.Datas.deletedatas_relations — Function

deletedatas_relations(identifier::String, headers::Dict, body::Dict, apiTest::Bool=false)

Suppression des relations sur une donnée.

exemple

Nakala.Datas.deletedatas_rights — Function

deletedatas_rights(identifier::String, headers::Dict, body::Dict, apiTest::Bool=false)

Suppression des droits pour un utilisateur ou un groupe d'utilisateurs sur une donnée.

exemple

Nakala.Datas.deletedatas_uploads — Function

deletedatas_uploads(fileIdentifier::String, headers::Dict, apiTest::Bool=false)

Suppression d'un fichier déposé dans l'espace temporaire

exemple

Nakala.Datas.downloaddatas_files — Function

downloaddatas_files(identifier::String, outputDir::String, header::Dict, apiTest::Bool=false)

Téléchargement des fichiers déposés associés à une donnée.

exemple

Nakala.Datas.getdatas — Function

getdatas(identifier::String, headers::Dict, apiTest=false)

Récupération des informations sur une donnée.

exemple

Nakala.Datas.getdatas_collections — Function

getdatas_collections(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération de la liste des collections contenant la donnée.

exemple

Nakala.Datas.getdatas_files — Function

getdatas_files(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération des métadonnées des fichiers associés à une donnée.

exemple

Nakala.Datas.getdatas_metadatas — Function

getdatas_metadatas(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération de la liste des métadonnées d'une donnée.

exemple

Nakala.Datas.getdatas_relations — Function

getdatas_relations(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération de la liste des relations d'une donnée.

exemple

Nakala.Datas.getdatas_rights — Function

getdatas_rights(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération des groupes et des utilisateurs ayant des droits sur la donnée.

exemple

Nakala.Datas.getdatas_status — Function

getdatas_status(identifier::String, headers::Dict, apiTest::Bool=false)

Récupération du statut d'une donnée.

exemple

Nakala.Datas.getdatas_uploads — Function

getdatas_uploads(headers::Dict, apiTest::Bool=false)

Récupération pour un utilisateur de la liste des objets fichiers déposés non encore associés à une donnée.

exemple

Nakala.Datas.getdatas_version — Function

getdatas_version(identifier::String, headers::Dict, apiTest=false)

Récupération de la liste des versions d'une donnée.

exemple

Nakala.Datas.postdatas — Function

postdatas(headers::Dict, body::Dict, apiTest=false)

Dépôt d'une nouvelle donnée.

exemple

Nakala.Datas.postdatas_collections — Function

postdatas_collections(identifier::String, headers::Dict, body::Array, apiTest::Bool=false)

Ajout d'une donnée dans un ensemble de collections.

exemple

Nakala.Datas.postdatas_files — Function

postdatas_files(identifier::String, headers::Dict, body::Dict, apiTest::Bool=false)

Ajout d'un fichier à une donnée.

exemple

Nakala.Datas.postdatas_metadatas — Function

postdatas_metadatas(identifier::String, headers::Dict, body::Dict, apiTest::Bool=false)

Suppression de métadonnées pour une donnée.

exemple

Nakala.Datas.postdatas_relations — Function

postdatas_relations(identifier::String, headers::Dict, body::Array, apiTest::Bool=false)

Ajout de relations sur une donnée.

exemple

Nakala.Datas.postdatas_rights — Function

postdatas_rights(identifier::String, headers::Dict, body::Array, apiTest::Bool=false)

Ajout de droits sur une donnée.

exemple

Nakala.Datas.postdatas_uploads — Function

postdatas_uploads(file::String, headers::Dict, apiTest::Bool=false)

Dépôt de fichier.

exemple

Nakala.Datas.putdatas — Function

putdatas(identifier::String, headers::Dict, body::Dict, apiTest=false)

Modification des informations d'une donnée.

exemple

Nakala.Datas.putdatas_collections — Function

putdatas_collections(identifier::String, headers::Dict, body::Array, apiTest::Bool=false)

Remplacement de l'ensemble des collections d'une donnée.

exemple

Nakala.Datas.putdatas_status — Function

putdatas_status(identifier::String, status::String, headers::Dict, apiTest::Bool=false)

Changement du statut d'une donnée.

exemple